[FIVE-125] MM:
- Creado el componente ManageCategories con el código común del selector de categorías en los componentes NewProjectDialog y EditProject.
- Movida la función fillProjectCategories al componente ProjectsList para recibirla por props en ambos componente mencionados, en lugar de duplicar el código en cada uno.
